refactor: Remove custom root namespace from Menu
