+++ Source/Menu/KeyInputControl.cs
@@ -20,7 +20,7 @@ using System.Drawing;
 using System.Windows.Forms;
 using ORTS.Settings;
 
-namespace ORTS
+namespace Menu
 {
     /// <summary>
     /// A control for viewing and altering keyboard input settings, in combination with <see cref="KeyInputEditControl"/>.
