+++ Source/Menu/Program.cs
@@ -21,7 +21,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Windows.Forms;
 
-namespace ORTS
+namespace Menu
 {
     static class Program
     {
