+++ Source/Menu/RDButtonInputControl.cs
@@ -21,7 +21,7 @@ using System.Windows.Forms;
 
 using ORTS.Common.Input;
 
-namespace Orts.Menu
+namespace Menu
 {
     /// <summary>
     /// A control for viewing and altering keyboard input settings, in combination with <see cref="KeyInputEditControl"/>.
