+++ Source/Menu/ResumeForm.cs
@@ -62,7 +62,7 @@ using ORTS.Menu;
 using ORTS.Settings;
 using Path = System.IO.Path;
 
-namespace ORTS
+namespace Menu
 {
     public partial class ResumeForm : Form
     {
