+++ Source/Menu/SortableBindingList.cs
@@ -22,7 +22,7 @@ using System.ComponentModel;
 using System.Linq;
 using System.Text;
 
-namespace ORTS
+namespace Menu
 {
     public class SortableBindingList<T> : BindingList<T>
     {
